Track the number of rounds played so far, fix #6
The problem was that the if statement would ignore the first block as soon as it saw that the boardState had no zeroes. Any game of tic-tac-toe will always consist of a maximum of 9 moves. If the game isn't at 9 moves yet, this check will fail and the code will default to checking the first if block properly